first smart enemy behavior

+++ LuMaxMan/GameplayConfiguration.swift
@@ -24,8 +24,8 @@ struct GameplayConfiguration {
     
     struct Enemy {
         // The static settings for enemies.
-        static let movementSpeedWhenFollowing: CGFloat = 230.0
-        static let movementSpeedWhenEscaping: CGFloat = 150.0
+        static let movementSpeedWhenFollowing: Float = 230.0
+        static let movementSpeedWhenEscaping: Float = 150.0
         // needed?
         static let appearDuration: NSTimeInterval = 0.5
         // needed?
